requirements.txt
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cftime==1.6.4.post1
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
docopt-ng==0.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.30.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lxml==6.0.2
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.3
matplotlib==3.9.4
narwhals==2.15.0
netCDF4==1.7.2
nltk==3.9.2
numpy==2.0.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
pillow==11.3.0
plotly==6.5.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydap==3.5
pydeck==0.9.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.36.0
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
streamlit==1.50.0
streamlit-autorefresh==1.0.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
WebOb==1.8.9
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
xarray==2024.7.0
xgboost==2.1.4
zipp==3.23.0
